add foreign key to notes owner column

+++ src/services/notes/controller/note-controller.js
@@ -1,11 +1,13 @@
 import NoteRepositories from '../repositories/note-repositories.js';
 import { InvariantError, NotFoundError } from '../../../exceptions/index.js';
 import response from '../../../utils/response.js';
+import AuthorizationError from '../../../exceptions/authorization-error.js';
 
 export const createNote = async (req, res, next) => {
   const { title, tags, body } = req.validated;
+  const { id: owner } = req.user;
 
-  const note = await NoteRepositories.createNote({ title, tags, body });
+  const note = await NoteRepositories.createNote({ title, tags, body, owner });
 
   if (!note) {
     return next(new InvariantError('Gagal menambahkan catatan'));
@@ -16,14 +18,37 @@ export const createNote = async (req, res, next) => {
   });
 };
 
-export const getAllNotes = async (req, res) => {
-  const notes = await NoteRepositories.getAllNotes();
-
-  return response(res, 200, 'Catatan sukses ditampilkan', { notes });
+export const getAllNotes = async (req, res, next) => {
+  try {
+    const { id: owner } = req.user;
+    const notes = await NoteRepositories.getNotes(owner);
+    const noteMap = notes.map((note) => ({
+      noteId:note.id,
+      title: note.title,
+      createdAt: note.created_at,
+      updatedAt: note.updated_at,
+      tags: note.tags,
+      body: note.body,
+    }));
+    return response(res, 200, 'Catatan sukses ditampilkan', { notes: noteMap });
+  } catch (err) {
+    next(err);
+  }
 };
+
+/*  */
 
 export const getNotesById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
+
   const note = await NoteRepositories.getNoteById(id);
 
   if (!note) {
@@ -35,7 +60,14 @@ export const getNotesById = async (req, res, next) => {
 
 export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
-  const { title, tags, body } = req.validated;
+  const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
 
   const note = await NoteRepositories.editNoteById({ id, title, tags, body });
 
@@ -48,11 +80,18 @@ export const editNoteById = async (req, res, next) => {
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
   const deletedNote = await NoteRepositories.deleteNoteById(id);
 
   if (!deletedNote) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
 
-  return response(res, 200, 'Catatan berhasil dihapus');
+  return response(res, 200, 'Catatan berhasil dihapus', null);
 };
